✨ Both extension and cli use the same runner.
Now the list of audits / gatherers is in one file, the only difference
is now the driver that is powering them.
